+++ CHANGELOG.md
@@ -11,5 +11,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 - Added code for class: SchemaOCDS
 - Added code for function: common_checks_ocds
+- Added code for function: ocds_json_output
+- Added CLI
+